agent/requirements_agent.txt
# GlassTrax Agent - Minimal Dependencies
# These are bundled with the standalone installer
# For development, use the main requirements.txt

# Core API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
ruamel.yaml>=0.18.0

# Authentication
bcrypt>=4.0.0

# Database (has pre-built wheels for win32)
pyodbc>=5.0.0

# System Tray (Windows)
pystray>=0.19.5
Pillow>=10.0.0
